% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gridsearch.R
\name{ds.gridsearch}
\alias{ds.gridsearch}
\title{Remote "gridsearch"}
\usage{
ds.gridsearch(m, minit, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{m, }{first argument to be sent to gridsearch on the remote nodes, a call to the hlme function. Please check the original documetation of the function gridsearch.
The call should be protected by quote(...) to protect from early, local evaluation. Ex: gridsearch(m = quote(hlme(...)), mint = 'obj', ...)}

\item{minit, }{a character, the name of the object on the server that was created with the newObj argument of a previous call of the function ds.hlme.
pPlease check the documentation of that function.}

\item{..., }{remainder of the gridsearch arguments. It is recommended to pass all the arguments by name.}

\item{async}{same as in datashield.aggregate}

\item{datasources}{same as in datashield.aggregate}
}
\description{
Executes "gridsearch" from the library "lcmm" on the remote nodes
}
