% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.jlcmm.R
\name{ds.jlcmm}
\alias{ds.jlcmm}
\title{Remote "jlcmm"}
\usage{
ds.jlcmm(data, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{data, }{first argument to be sent to jlcmm on the remote nodes. Please check the original function documentation for details. data should be a character, enclosed in quotes.}

\item{..., }{remainder of the jlcmm arguments.}

\item{async}{same as in datashield.aggregate}

\item{datasources}{same as in datashield.aggregate}
}
\description{
Executes "jlcmm" from the library "lcmmm" on the remote nodes
}
