% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lcmm.R
\name{ds.lcmm}
\alias{ds.lcmm}
\title{Remote "lcmm"}
\usage{
ds.lcmm(data, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{data, }{first argument to be sent to lcmm on the remote nodes. Please check the original function documentation for details. data should be a character, enclosed in quotes.}

\item{..., }{remainder of the lcmm arguments.}

\item{async}{same as in datashield.aggregate}

\item{datasources}{same as in datashield.aggregate}
}
\description{
Executes "lcmm" from the library "lcmmm" on the remote nodes
}
