% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.mpjlcmm.R
\name{ds.mpjlcmm}
\alias{ds.mpjlcmm}
\title{Remote "mpjlcmm"}
\usage{
ds.mpjlcmm(data, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{data, }{first argument to be sent to mpjlcmm on the remote nodes. Please check the original function documentation for details. data should be a character, enclosed in quotes.}

\item{..., }{remainder of the mpjlcmm arguments.}

\item{async}{same as in datashield.aggregate}

\item{datasources}{same as in datashield.aggregate}
}
\description{
Executes "mpjlcmm" from the library "lcmmm" on the remote nodes
}
