% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.multlcmm.R
\name{ds.multlcmm}
\alias{ds.multlcmm}
\title{Remote "multlcmm"}
\usage{
ds.multlcmm(data, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{data, }{first argument to be sent to multlcmm on the remote nodes. Please check the original function documentation for details. data should be a character, enclosed in quotes.}

\item{..., }{remainder of the multlcmm arguments.}

\item{async}{same as in datashield.aggregate}

\item{datasources}{same as in datashield.aggregate}
}
\description{
Executes "multlcmm" from the library "lcmmm" on the remote nodes
}
