% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.hlme.R
\name{ds.hlme}
\alias{ds.hlme}
\title{Remote "hlme"}
\usage{
ds.hlme(data, ..., async = TRUE, datasources = NULL)
}
\arguments{
\item{data, }{first argument to be sent to hlme on the remote nodes. Please check the original function documentation for details. data should be a character, enclosed in quotes.}

\item{..., }{remainder of the hlme arguments.}

\item{async}{same as in datashield.aggregate}

\item{datasources}{same as in datashield.aggregate}
}
\description{
Executes "hlme" from the library "lcmmm" on the remote nodes
}
